+++ test/test_aqua.jl
@@ -1,4 +1,4 @@
 using Aqua
 using ResumableFunctions
 
-Aqua.test_all(ResumableFunctions)
+Aqua.test_all(ResumableFunctions; deps_compat = (; check_weakdeps = false))